Move to static methods
